+++ trixnity-core/src/commonMain/kotlin/net/folivo/trixnity/core/model/events/block/m/TopicContentBlock.kt
@@ -1,0 +1,23 @@
+package net.folivo.trixnity.core.model.events.block.m
+
+import kotlinx.serialization.Contextual
+import kotlinx.serialization.Serializable
+import net.folivo.trixnity.core.model.events.block.EventContentBlock
+import net.folivo.trixnity.core.model.events.block.EventContentBlocks
+import kotlin.jvm.JvmInline
+
+@Serializable
+@JvmInline
+value class TopicContentBlock(
+    val blocks: @Contextual EventContentBlocks,
+) : EventContentBlock.Default {
+    constructor(text: TextContentBlock) : this(EventContentBlocks(text))
+
+    override val type: EventContentBlock.Type<TopicContentBlock> get() = Type
+
+    companion object Type : EventContentBlock.Type<TopicContentBlock> {
+        override val value: String = "m.topic"
+    }
+
+    val text: TextContentBlock? get() = blocks[TextContentBlock]
+}